📝: docker-compose.yaml 생성, flask 서버 생성
