+++ ai/Data/example.json
@@ -2,7 +2,7 @@
     "noteId": 1,
     "title": "title",
     "folder": "folder",
-    "createdAt": "2024-11-25 19:26:55",
+    "createdAt": "2024-11-25 22:10:20",
     "duration": "00:23:15",
     "content": {
         "script": [
