+++ ai/dockerfile
@@ -1,0 +1,15 @@
+# syntax=docker/dockerfile:1
+FROM python:3.9-slim
+
+WORKDIR /app
+
+# 필요한 파일 복사
+COPY requirement.txt ./
+RUN pip install --no-cache-dir -r requirement.txt
+
+# 애플리케이션 파일 복사
+COPY . .
+
+EXPOSE 5000
+
+CMD ["python", "app.py"]
