+++ ai/requirement.txt
@@ -1,0 +1,3 @@
+flask
+requests
+python-dotenv
