+++ docker-compose.yaml
@@ -1,0 +1,15 @@
+version: '3.8'
+services:
+  ai-server:
+    build:
+      context: ./ai       # 빌드 컨텍스트를 ai 디렉토리로 지정
+    ports:
+      - "5000:5000"
+    networks:
+      - bento-network
+    env_file:
+      - .env
+
+networks:
+  bento-network:
+    external: true  # 외부 네트워크 사용
